Move worker types into server package
